Factories and mines adapt to market pressures and increase or reduce their production based on supply and demand. (#61)
* Some GDP work.
Calculate GDP per 25 hours based on how much spending is needed. Can probably calculate GDP per year with these metrics.
* Some tweaking of the economy to balance it.
Population consumes about the amount of an average American
Enough resources are generated so that we can satisfy the consumption needs of Americans. But it doesn't generate enough for population growth.

* Factories adjust their productivity.
Factories in theory generate less resources over time if there is an excess of resources on the market, and generate more if there is a lacking of resources on the market.

* Indicators and extra safeties
Indicator to show how much production there is.
Checks in economy to ensure there is no divide by zero errors.

* Change UI organization, change approach to mining
The amount of resources in miners are not generated based upon a generation resource ledger, and are now instead set in their productivity.
Change up market information tooltip so that supply, demand, and SD ratio are showed in one unified table.
Remove supply from resource ledger because we calculate that from the market resource ledger in the first place.